Fix garbage collection on windows
File paths added to the reference map during the download phase do not
match canonicalized file paths used during garbage collection. This can
be reproduced by running on windows with the default mirror and orig
directories, causing all files associated with the stable channel to be
garbage collected.
The canonicalize function requires the file to exist, hence the
introduction of the new normalize_path function.